backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0

pydantic>=2.8,<3
pydantic-settings>=2.2,<3

httpx==0.26.0
python-dotenv==1.0.1
python-multipart==0.0.6
# Database
sqlalchemy>=2.0,<3
psycopg2-binary==2.9.9
alembic==1.12.1

# Caching
redis==5.0.1
hiredis==2.2.3

# Web3
web3==6.11.3

# HTTP & Data
requests==2.31.0
pycoingecko==3.1.0

# Utils
cors==1.0.1

# Authentication
bcrypt==4.1.2
PyJWT==2.8.0